+++ page/src/app/JournalMainPage.jsx
