+++ page/src/app/config/urls.js
@@ -1,0 +1,3 @@
+// export constant backend_url 'http://127.0.0.1:5000/submit_journal';
+
+export const backend_url = 'http://127.0.0.1:5000/submit_journal';
